[frontend] ExternalLinkPopover generalization (#3403)
